+++ game-residentevil42023/index.js
@@ -60,6 +60,12 @@ const FLUFFYMOD_PATH = path.join("Games", FLUFFY_FOLDER, "Mods");
 const FLUFFYMOD_FILE = "modinfo.ini";
 const PAK_EXT = '.pak';
 
+const PRESET_ID = `${GAME_ID}-preset`;
+const PRESET_NAME = "Fluffy Preset";
+let PRESET_PATH = path.join("Games", FLUFFY_FOLDER, "Presets");
+const PRESET_PATH_DEMO = path.join("Games", FLUFFY_FOLDER_DEMO, "Presets");
+const PRESET_EXTS = [".prt"];
+
 const LOOSELUA_ID = `${GAME_ID}-looselua`;
 const LOOSELUA_NAME = "Loose Lua (REFramework)";
 const LOOSELUA_PATH = path.join(".");
@@ -258,6 +264,7 @@ function getExecutable(gamePath) {
   };
   if (isCorrectExec(EXEC_DEMO)) {
     MOD_PATH_USED = MOD_PATH_DEMO;
+    PRESET_PATH = PRESET_PATH_DEMO;
     return EXEC_DEMO; 
   };
   MOD_PATH_USED = MOD_PATH;
@@ -277,6 +284,7 @@ function getModPath(gamePath) {
   };
   if (isCorrectExec(EXEC_DEMO)) {
     MOD_PATH_USED = MOD_PATH_DEMO;
+    PRESET_PATH = PRESET_PATH_DEMO;
     return MOD_PATH_DEMO; 
   };
   MOD_PATH_USED = MOD_PATH;
@@ -726,6 +734,48 @@ function installUpscaler(files) {
   return Promise.resolve({ instructions });
 }
 
+//Installer test for Fluffy Preset files
+function testPreset(files, gameId) {
+  const isMod = files.some(file => PRESET_EXTS.includes(path.extname(file).toLowerCase()));
+  const isFluffy = files.some(file => path.basename(file).toLowerCase() === FLUFFYMOD_FILE);
+  let supported = (gameId === spec.game.id) && isMod && !isFluffy;
+
+  // Test for a mod installer
+  if (supported && files.find(file =>
+      (path.basename(file).toLowerCase() === 'moduleconfig.xml') &&
+      (path.basename(path.dirname(file)).toLowerCase() === 'fomod'))) {
+    supported = false;
+  }
+
+  return Promise.resolve({
+    supported,
+    requiredFiles: [],
+  });
+}
+
+//Install Fluffy Preset files
+function installPreset(files) {
+  const modFile = files.find(file => PRESET_EXTS.includes(path.extname(file).toLowerCase()));
+  const idx = modFile.indexOf(path.basename(modFile));
+  const rootPath = path.dirname(modFile);
+  const setModTypeInstruction = { type: 'setmodtype', value: PRESET_ID };
+
+  // Remove directories and anything that isn't in the rootPath.
+  const filtered = files.filter(file =>
+    ((file.indexOf(rootPath) !== -1) &&
+      (!file.endsWith(path.sep))));
+
+  const instructions = filtered.map(file => {
+    return {
+      type: 'copy',
+      source: file,
+      destination: path.join(file.substr(idx)),
+    };
+  });
+  instructions.push(setModTypeInstruction);
+  return Promise.resolve({ instructions });
+}
+
 //test for zips
 async function testZipContent(files, gameId) {
   const isFluffy = files.some(file => path.basename(file).toLowerCase() === FLUFFY_EXEC);
@@ -877,6 +927,7 @@ async function setup(discovery, api, gameSpec) {
   await downloadFluffy(api, gameSpec);
   await downloadREFramework(api, gameSpec);
   await fs.ensureDirWritableAsync(path.join(GAME_PATH, UPSCALER_PATH));
+  await fs.ensureDirWritableAsync(path.join(GAME_PATH, PRESET_PATH));
   return fs.ensureDirWritableAsync(path.join(GAME_PATH, MOD_PATH_USED));
 }
 
@@ -910,11 +961,22 @@ function applyGame(context, gameSpec) {
       var _a;
       return (gameId === GAME_ID) && !!((_a = context.api.getState().settings.gameMode.discovered[gameId]) === null || _a === void 0 ? void 0 : _a.path);
     }, 
-    (game) => pathPattern(context.api, game, path.join('{gamePath}', FLUFFYMOD_PATH)), 
+    (game) => pathPattern(context.api, game, path.join('{gamePath}', MOD_PATH_USED)), 
     () => Promise.resolve(false), 
     { 
       name: FLUFFYMOD_NAME, 
-      mergeMods: false,
+      mergeMods: reZip,
+    }
+  );
+  context.registerModType(PRESET_ID, 40, 
+    (gameId) => {
+      var _a;
+      return (gameId === GAME_ID) && !!((_a = context.api.getState().settings.gameMode.discovered[gameId]) === null || _a === void 0 ? void 0 : _a.path);
+    }, 
+    (game) => pathPattern(context.api, game, path.join('{gamePath}', PRESET_PATH)), 
+    () => Promise.resolve(false), 
+    { 
+      name: PRESET_NAME, 
     }
   );
 
@@ -924,6 +986,7 @@ function applyGame(context, gameSpec) {
   context.registerInstaller(LOOSELUA_ID, 29, testLooseLua, installLooseLua);
   context.registerInstaller(ROOT_ID, 31, testRoot, installRoot);
   context.registerInstaller(UPSCALER_ID, 33, testUpscaler, installUpscaler);
+  context.registerInstaller(PRESET_ID, 35, testPreset, installPreset);
   if (!reZip) {
     context.registerInstaller(FLUFFYMOD_ID, 45, testFluffyMod, installFluffyMod);
   } else {